refactor: restructure video player components; introduce PlayerControls, VideoError, and VideoLoading; implement useVideoSources hook for video source management
