+++ src/components/ui/video-player/VideoLoading.tsx
@@ -1,0 +1,7 @@
+export function VideoLoading() {
+	return (
+		<div className='aspect-video relative -top-full bg-field border border-border rounded-md animate-pulse flex items-center justify-center'>
+			<p className='text-lg text-muted-foreground'>Video loading...</p>
+		</div>
+	)
+}
